Changed the build system so that only one library is build. This reverts the previous change where PoDoFo was split into two libraries podofo-base and podofo-doc. Now, there is only one library libpodofo.
I reverted the change because of major linkage problems, which caused for example crashes in PdfMemDocument. This and other reasons kept many people from switching to trunk and to the latest version of PoDoFo.
These problem should all be gone and if you refused to upgrade, it should be safe now.

This is the basis for the upcoming PoDoFo 0.9.0 release, which I intend to release next week.