Patch by aki009
The tools are great, but can't handle standard input (or output) by default. I've attached a diff that implements this functionality for podofoxmp. The same could be copied by some enterprising soul to all the tools. Since quite a bit of the code is boiler plate, ideally the implementation would go into the library itself. It'd be great if the parser would know how to handle a redirected cin, as currently it uses seeks that aren't available for cin. (E.g. if the stream does not support seeks, the parser would have the option to fall back to a buffered mode.)