FIXED: Patch by Amin Massad to fix loading PDFs with object streams:
I have found and (hopefully) fixed a bug in PdfObjectStreamParserObject::ReadObjectsFromStream() which yielded a parse error for some PDFs with ObjectStreams. Could you please check the proposed patch and add it to the recent SVN version of PoDoFo? Thanks a lot!

Here comes the patch against SVN rev 1507:

Index: podofo-src-r1507/src/base/PdfObjectStreamParserObject.cpp
===================================================================
--- podofo-src-r1507/src/base/PdfObjectStreamParserObject.cpp	(revision 5315)
+++ podofo-src-r1507/src/base/PdfObjectStreamParserObject.cpp	(working copy)
@@ -105,6 +105,7 @@
 		}
 
         // move back to the position inside of the table of contents
+        device.Device()->Clear();
         device.Device()->Seek( pos );
 
         ++i;

The background of this patch is as follows.

SYMPTOMS: 
Got a PDF created by Adobe Illustrator 10 which yielded a parse error in PoDoFo, e.g. calling podofopdfinfo yielded the following output:

./build/bin/podofopdfinfo sample.pdf
Error: An error 5 ocurred during uncompressing the pdf file.


PoDoFo encounter an error. Error: 5 ePdfError_UnexpectedEOF
	Error Description: End of file was reached unxexpectedly.
	Callstack:
	#0 Error Source: podofo-svn/podofo-src/src/base/PdfParser.cpp:213
		Information: Unable to load objects from file.
	#1 Error Source: podofo-svn/podofo-src/src/base/PdfTokenizer.cpp:340
		Information: Expected number

CAUSE: 
The PDF contains an ObjectStream which ends where the last three objects are of type "number". The *END* of the object stream looks like this:
18272 19199 11818

There is a problem in the function PdfObjectStreamParserObject::ReadObjectsFromStream() because it uses *TWO* tokenizers (line 75 and line 89) on *ONE* InputDevice "device". Now, the call of variantTokenizer.GetNextVariant() in line 90 for the third last number, i.e. "18272" in the example above, trigger EOF on the device because GetNextToken() always reads three tokens. 

After that we want to "move back to the position inside of the table of contents" (line 107f) by calling device.Device()->Seek( pos ). This does not work anymore because EOF was already reached and reading the next object number lObj in line 81 from the TOC of the object stream fails. Hence, we get an error before we cab reach the second last number "19199" of the object stream.

72 void PdfObjectStreamParserObject::ReadObjectsFromStream( char* pBuffer, pdf_long lBufferLen, long long lNum, long long lFirst, ObjectIdList const & list)
     73 {
     74     PdfRefCountedInputDevice device( pBuffer, lBufferLen );
     75     PdfTokenizer             tokenizer( device, m_buffer );
     76     PdfVariant               var;
     77     int                      i = 0;
     78 
     79     while( static_cast<long long>(i) < lNum )
     80     {
     81         const long long lObj     = tokenizer.GetNextNumber();
     82         const long long lOff     = tokenizer.GetNextNumber();
     83         const std::streamoff pos = device.Device()->Tell();
     84 
     85         // move to the position of the object in the stream
     86         device.Device()->Seek( static_cast<std::streamoff>(lFirst + lOff) );
     87 
     88                 // use a second tokenizer here so that anything that gets dequeued isn't left in the tokenizer that reads the offsets and lengths
     89             PdfTokenizer variantTokenizer( device, m_buffer );
     90         variantTokenizer.GetNextVariant( var, m_pEncrypt );
     91                 bool should_read = std::find(list.begin(), list.end(), lObj) != list.end();
     92 #if defined(PODOFO_VERBOSE_DEBUG)
     93         std::cerr << "ReadObjectsFromStream STREAM=" << m_pParser->Reference().ToString() <<
     94                         ", OBJ=" << lObj <<
     95                         ", " << (should_read ? "read" : "skipped") << std::endl;
     96 #endif
     97                 if (should_read)
     98         {
     99                         if(m_vecObjects->GetObject(PdfReference( static_cast<int>(lObj), 0LL ))) 
    100             {
    101                 PdfError::LogMessage( eLogSeverity_Warning, "Object: %li 0 R will be deleted and loaded again.\n", lObj );
    102                 delete m_vecObjects->RemoveObject(PdfReference( static_cast<int>(lObj), 0LL ),false);
    103             }
    104             m_vecObjects->insert_sorted( new PdfObject( PdfReference( static_cast<int>(lObj), 0LL ), var ) );
    105                 }
    106 
    107         // move back to the position inside of the table of contents
    108         device.Device()->Clear(); // ** NEWLY ADDED ** 
    109         device.Device()->Seek( pos );
    110 
    111         ++i;
    112     }
    113 }

SOLUTION:
To clear EOF, I have add line 108 to solve the parse error (clearing only std::ios_base::eofbit was not enough, failbit seems to be a problem, too). Now, the inner tokenizer does not break the reading of the TOC and the PDF is parsed normally to the end.

Best regards,
Amin