Implement incremental update for PdfMemDocument
The Load() methods received a new bForUpdate argument,
which is used to determine whether to load the document
for incremental update. The default value is 'false'.
The documents opened for incremental update can be saved
using WriteUpdate(), which copies the original content
first and then writes the update at the end of the file.
In case a real file had been loaded, and the same overload
WriteUpdate() method is used (const char * for const char *,
const wchar_t * for const wchar_t *) and the file names
match, then the file is not rewritten completely, but only
the changes are appended at the end of it. In all other
cases a write to a new stream/file is required, due to
the file content copy from the source file to the destination
file (when the copy is done, the destination file is truncated
first, but only then the copy operation is initiated).