Patch by Chris: podofoimpose: Properly copy 'number' and 'null' objects
In the /FontFile in the input document the /Length1, /Length2 and /Length3 had
a reference to 'null' or a 'number' value. With the previous method of podofoimpose
those references got messed up in the output document (wrong ref). The fix writes
the correct references to 'number' and 'null' data types for the output document.